% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLAWexplorer.R
\name{plot_features.SLAWexplorer}
\alias{plot_features.SLAWexplorer}
\title{Plot feature across the selected files}
\usage{
\method{plot_features}{SLAWexplorer}(
  slexp,
  fids,
  mz_margin = 0.005,
  rt_margin = 0.05,
  layout_matrix = NULL,
  ...
)
}
\arguments{
\item{features}{}
}
\value{

}
\description{
Plot feature across the selected files
}
\examples{
print("Examples to be put here.")
}
