% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLAWexplorer.R
\name{plot_peaks.SLAWexplorer}
\alias{plot_peaks.SLAWexplorer}
\title{Title}
\usage{
\method{plot_peaks}{SLAWexplorer}(slexp, mzlim, rtlim, colors, ...)
}
\arguments{
\item{slexp}{A SLAWexplorer object}

\item{mzlim}{An mz limit}

\item{rtlim}{A rtlimit}
}
\value{
None
}
\description{
Title
}
\examples{
print("Examples to be put here")
}
