% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLAWexplorer.R
\name{select_samples.SLAWexplorer}
\alias{select_samples.SLAWexplorer}
\title{Select samples from a SLAWexplorer acquisition}
\usage{
\method{select_samples}{SLAWexplorer}(slexp, samples)
}
\arguments{
\item{slexp}{A SLAWexplorer object}

\item{samples}{A set of samples id as integer (Their position in the columns)}
}
\value{

}
\description{
Select samples from a SLAWexplorer acquisition
}
\examples{
print("Examples to be put here")
}
