% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLAWexplorer.R
\name{SLAWexplorer}
\alias{SLAWexplorer}
\title{SLAWexplorer TO Visualise SLAW Results}
\usage{
SLAWexplorer(path, raw_path)
}
\arguments{
\item{path}{The path of the SLAW output directory}

\item{raw_path}{The path of the rawe files for EICs visualisation}
}
\value{
A SLAWexplorer object
}
\description{
SLAWexplorer TO Visualise SLAW Results
}
\examples{
print("Examples to be put here")
}
