% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLAWvis.R
\docType{package}
\name{SLAWvis}
\alias{SLAWvis}
\title{The SLAW visualisation package}
\description{
SLAWvis is a visualisation pakcage which is designed ot be used in with the containerized version of SLAW for the data
}
